Refactor: Start using feature-first architecture
It's a goal to refactor the source to have a clean feature-first
architecture. Start by putting the new sharing feature into a separate
folder features/share/
Increase test coverage for ShareButton to 100%
