+++ test/view_page_test.dart
@@ -2,10 +2,10 @@ import 'package:app4training/data/app_language.dart';
 import 'package:app4training/data/exceptions.dart';
 import 'package:app4training/data/globals.dart';
 import 'package:app4training/data/languages.dart';
+import 'package:app4training/features/share/share_button.dart';
 import 'package:app4training/l10n/l10n.dart';
 import 'package:app4training/routes/view_page.dart';
 import 'package:app4training/widgets/language_selection_button.dart';
-import 'package:app4training/widgets/share_button.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
